# doctolib/swagger/swaggerDoctolib.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctolib Management",
        "description": "Doctolib Management",
        "version": "V1"
    },
    "paths": {
...
        "/patients": {
            "get": {
                "tags": [
                    "patients"
                ],
                "summary": "Returns a map of all patient or a map of patient according to a designated lastname",
                "description": "Returns a map of all patient or a map of patient according to a designated lastname",
                "operationId": "getPatients",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientDTO"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Please contact us"
                    },
                    "404": {
                        "description": "No patients found"
                    }
                }
            },
            "post": {
                "tags": [
                    "patients"
                ],
                "summary": "Create a patient",
                "description": "Create a patient",
                "operationId": "createPatient",
                "responses": {
                    "201": {
                        "description": "Successful operation but no content"
                    },
                    "500": {
                        "description": "Please contact us"
                    }
                }
            }
        },
        "/patients/{id}": {
            "put": {
                "tags": [
                    "patients"
                ],
                "summary": "Update a patient according to ID",
                "description": "Update a patient according to ID",
                "operationId": "updatePatient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the patient that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorDTO"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Please contact us"
                    }
                }
            },
            "delete": {
                "tags": [
                    "patients"
                ],
                "summary": "Delete a patient according to ID",
                "description": "Delete a patient according to ID",
                "operationId": "deletePatients",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the patient that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation but no content"
                    },
                    "500": {
                        "description": "Please contact us"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "DoctorDTO": {
                "title": "DoctorDTO model",
                "description": "DoctorDTO model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "integer"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "format": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string",
                        "format": "string"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "lastName",
                        "type": "string",
                        "format": "string"
                    },
                    "firstName": {
                        "title": "firstName",
                        "description": "firstName",
                        "type": "string",
                        "format": "string"
                    },
                    "specialty": {
                        "title": "specialty",
                        "description": "specialty",
                        "type": "string",
                        "format": "string"
                    },
                    "addressNum": {
                        "title": "addressNum",
                        "description": "addressNum",
                        "type": "integer",
                        "format": "integer"
                    },
                    "addressStreet": {
                        "title": "addressStreet",
                        "description": "addressStreet",
                        "type": "string",
                        "format": "string"
                    },
                    "postalCodeAddress": {
                        "title": "postalCodeAddress",
                        "description": "postalCodeAddress",
                        "type": "integer",
                        "format": "integer"
                    },
                    "townAddress": {
                        "title": "townAddress",
                        "description": "townAddress",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object"
            },
            "PatientDTO": {
                "title": "PatientDTO model",
                "description": "PatientDTO model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "integer"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "format": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string",
                        "format": "string"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "lastName",
                        "type": "string",
                        "format": "string"
                    },
                    "firstName": {
                        "title": "firstName",
                        "description": "firstName",
                        "type": "string",
                        "format": "string"
                    },
                    "birthDate": {
                        "title": "birthDate",
                        "description": "birthDate",
                        "type": "string",
                        "format": "datetime"
                    },
                    "addressNum": {
                        "title": "addressNum",
                        "description": "addressNum",
                        "type": "integer",
                        "format": "integer"
                    },
                    "addressStreet": {
                        "title": "addressStreet",
                        "description": "addressStreet",
                        "type": "string",
                        "format": "string"
                    },
                    "postalCodeAddress": {
                        "title": "postalCodeAddress",
                        "description": "postalCodeAddress",
                        "type": "integer",
                        "format": "integer"
                    },
                    "townAddress": {
                        "title": "townAddress",
                        "description": "townAddress",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}